newsletter_scraper/requirements.txt
pymongo==4.6.0
requests==2.31.0
beautifulsoup4==4.12.2
imaplib2==3.6
spacy==3.7.2
validators==0.22.0
APScheduler==3.10.4
python-dotenv==1.0.0
openai==1.3.7
tldextract==5.1.1
urllib3==2.1.0
lxml==4.9.3
